Update HPO resources and create Makefile to automatically create them.
