--- data/makefiles/Makefile_ABB-SCORE
@@ -1,11 +1,0 @@
-help:
-	@cat Makefile_ABB-SCORE
-
-download:
-	wget -c https://public_docs.crg.es/sossowski/publication_data/ABB/ABB_SCORE.txt
-
-convert:
-	python3 create_ABB-SCORE_bed.py ABB_SCORE.txt hg19_ABB-SCORE.bed
-	bgzip hg19_ABB-SCORE.bed
-	tabix -p bed hg19_ABB-SCORE.bed.gz
-	rm ABB_SCORE.txt
